feat: Adding alert dialog confirm to delete song
Issue: https://github.com/GeovaniTech/ListenNow-App/issues/77

+++ app/src/main/java/br/com/listennow/fragments/HomeFragment.kt
@@ -278,7 +278,13 @@ class HomeFragment : CommonFragment<HomeViewModel, FragmentHomeBinding>(), ICont
         }
 
         dataBinding.deleteSongButton.setOnClickListener {
-            viewModel.deleteSong(item)
+            showAlertDialog(
+                title = getString(R.string.are_you_sure),
+                message = getString(R.string.home_are_you_sure_about_deleting_song, item.name),
+                positiveClick = {
+                    viewModel.deleteSong(item)
+                }
+            )
         }
     }
